refactor(ui): extract types and pure utils to domain and utils files (Phase 1)
